Add files via upload
The University Placement Portal is an all-encompassing platform that serves as a pivotal tool for facilitating seamless communication and coordination between the Training and Placement Office (TPO), students, and companies. This comprehensive portal offers a range of features designed to streamline the placement process and enhance the overall experience for all stakeholders involved.

One of the key functionalities of the portal is the ability for the TPO to add student profiles. This feature allows for the capturing of important details such as academic achievements, skills, project work, and other relevant information. By having a comprehensive profile, students can be efficiently matched with suitable job opportunities based on their qualifications and interests. This not only benefits the students in finding the right placement but also aids companies in identifying potential candidates.

In addition to student profiles, the TPO can also add company profiles to the portal. These profiles provide detailed information about participating organizations, including their industry, job roles,